Create a specific graphic render for MessageController
